+++ BabelWiresLib/Processors/processor.cpp
@@ -7,7 +7,37 @@
  **/
 #include <BabelWiresLib/Processors/processor.hpp>
 
+#include <BabelWiresLib/Features/modelExceptions.hpp>
+#include <BabelWiresLib/Features/rootFeature.hpp>
+#include <BabelWiresLib/Features/simpleValueFeature.hpp>
+#include <BabelWiresLib/Project/projectContext.hpp>
+#include <BabelWiresLib/TypeSystem/typeRef.hpp>
+
+#include <Common/Identifiers/registeredIdentifier.hpp>
+
+babelwires::Processor::Processor(const ProjectContext& projectContext, const TypeRef& inputTypeRef,
+                                 const TypeRef& outputTypeRef)
+    : m_inputFeature(std::make_unique<babelwires::SimpleValueFeature>(projectContext.m_typeSystem, inputTypeRef))
+    , m_outputFeature(std::make_unique<babelwires::SimpleValueFeature>(projectContext.m_typeSystem, outputTypeRef)) {
+    const Type* const inputType = inputTypeRef.tryResolve(projectContext.m_typeSystem);
+    if (!inputType) {
+        throw ModelException() << "Input type reference " << inputTypeRef << " could not be resolved";
+    }
+    const Type* const outputType = outputTypeRef.tryResolve(projectContext.m_typeSystem);
+    if (!outputType) {
+        throw ModelException() << "Output type reference " << outputTypeRef << " could not be resolved";
+    }
+}
+
 babelwires::Processor::~Processor() = default;
+
+babelwires::ValueFeature* babelwires::Processor::getInputFeature() {
+    return m_inputFeature.get();
+}
+
+babelwires::ValueFeature* babelwires::Processor::getOutputFeature() {
+    return m_outputFeature.get();
+}
 
 const babelwires::ValueFeature* babelwires::Processor::getInputFeature() const {
     return const_cast<Processor*>(this)->getInputFeature();
@@ -15,4 +45,9 @@ const babelwires::ValueFeature* babelwires::Processor::getInputFeature() const {
 
 const babelwires::ValueFeature* babelwires::Processor::getOutputFeature() const {
     return const_cast<Processor*>(this)->getOutputFeature();
+}
+
+void babelwires::Processor::process(UserLogger& userLogger) {
+    BackupScope scope(*m_outputFeature);
+    processValue(userLogger, *m_inputFeature, *m_outputFeature);
 }
