+++ BabelWiresLib/Processors/valueProcessor.hpp
@@ -23,18 +23,5 @@ namespace babelwires {
     class ValueProcessor : public Processor {
       public:
         ValueProcessor(const ProjectContext& projectContext, const TypeRef& inputTypeRef, const TypeRef& outputTypeRef);
-
-      protected:
-        /// Note: Implementations do not need to worry about backing-up or resolving changes in the output feature.
-        virtual void processValue(UserLogger& userLogger, const ValueFeature& inputFeature,
-                                  ValueFeature& outputFeature) const = 0;
-      public:
-        ValueFeature* getInputFeature() override;
-        ValueFeature* getOutputFeature() override;
-        void process(UserLogger& userLogger) override final;
-
-      protected:
-        std::unique_ptr<SimpleValueFeature> m_inputFeature;
-        std::unique_ptr<SimpleValueFeature> m_outputFeature;
     };
 } // namespace babelwires
